Some fixes based on the code review
